Split layout visualizer classes in separate files
- Make new story_of_your_layout directory inside inspector/flutter
