+++ packages/devtools_app/lib/src/inspector/flutter/story_of_your_layout/utils.dart
@@ -1,0 +1,62 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter/widgets.dart';
+
+///      | top    |
+/// left | child  | right
+///      | bottom |
+@immutable
+class GridAddOns extends StatelessWidget {
+  const GridAddOns({
+                     Key key,
+                     this.left,
+                     this.top,
+                     this.right,
+                     this.bottom,
+                     @required this.child,
+                   })  : assert(child != null),
+      super(key: key);
+
+  final Widget child;
+  final Widget top;
+  final Widget left;
+  final Widget right;
+  final Widget bottom;
+
+  CrossAxisAlignment get crossAxisAlignment {
+    if (left != null && right != null) {
+      return CrossAxisAlignment.center;
+    } else if (left == null && right != null) {
+      return CrossAxisAlignment.start;
+    } else if (left != null && right == null) {
+      return CrossAxisAlignment.end;
+    } else {
+      return CrossAxisAlignment.start;
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    print(crossAxisAlignment);
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      mainAxisAlignment: MainAxisAlignment.start,
+      crossAxisAlignment: crossAxisAlignment,
+      children: <Widget>[
+        if (top != null) top,
+        Row(
+          mainAxisSize: MainAxisSize.min,
+          children: <Widget>[
+            if (left != null) left,
+            Flexible(child: child),
+            if (right != null) right,
+          ],
+        ),
+        if (bottom != null) bottom,
+      ]);
+  }
+}
